+++ core/build.gradle
@@ -27,21 +27,18 @@ compileGroovy {
 gradlePlugin {
     plugins {
         binrayRelease {
-            id = "com.novoda.bintray-release"
-            implementationClass = "com.novoda.gradle.release.ReleasePlugin"
-        }
-        legacy {
-            id = "bintray-release"
-            implementationClass = "com.novoda.gradle.release.ReleasePlugin"
+            id = "guru.stefma.bintrayrelease"
+            implementationClass = "guru.stefma.bintrayrelease.ReleasePlugin"
         }
     }
 }
 
 publish {
-    userOrg = 'novoda'
-    groupId = 'com.novoda'
-    artifactId = rootProject.name
-    publishVersion = '0.8.1'
+    userOrg = 'stefma'
+    groupId = 'guru.stefma.bintrayrelease'
+    artifactId = "bintrayrelease"
+    uploadName = "BintrayRelease"
+    publishVersion = '1.0.0'
     desc = 'Super duper easy way to release your Android and other artifacts to bintray'
-    website = "https://github.com/novoda/${rootProject.name}"
+    website = "https://github.com/stefma/bintray-release"
 }
