--- core/src/main/groovy/com/novoda/gradle/release/JavaArtifacts.groovy
@@ -1,30 +1,0 @@
-package com.novoda.gradle.release
-
-import org.gradle.api.Project
-import org.gradle.api.tasks.bundling.Jar
-
-class JavaArtifacts {
-
-    def all(String publicationName, Project project) {
-        [sourcesJar(publicationName, project), javadocJar(publicationName, project)]
-    }
-
-    def sourcesJar(String publicationName, Project project) {
-        project.task(publicationName + 'SourcesJar', type: Jar) {
-            classifier = 'sources'
-            from project.sourceSets.main.allSource
-        }
-    }
-
-    def javadocJar(String publicationName, Project project) {
-        project.task(publicationName + 'JavadocJar', type: Jar) {
-            classifier = 'javadoc'
-            from project.javadoc.destinationDir
-        }
-    }
-
-    def from(Project project) {
-        project.components.java
-    }
-
-}
