+++ core/src/main/groovy/guru/stefma/bintrayrelease/AndroidPublishExtension.groovy
@@ -1,0 +1,18 @@
+package guru.stefma.bintrayrelease
+
+import guru.stefma.androidartifacts.AndroidArtifactsExtension
+
+class AndroidPublishExtension extends PublishExtension {
+
+    private AndroidArtifactsExtension mAndroidExtension
+
+    AndroidPublishExtension(AndroidArtifactsExtension androidArtifactsExtension) {
+        mAndroidExtension = androidArtifactsExtension
+    }
+
+    @Override
+    void setArtifactId(String newArtifactId) {
+        super.setArtifactId(newArtifactId)
+        mAndroidExtension.artifactId = newArtifactId
+    }
+}
