--- core/src/test/groovy/com/novoda/gradle/release/GradleVerionsParams.groovy
@@ -1,21 +1,0 @@
-package com.novoda.gradle.release
-
-import org.gradle.testkit.runner.TaskOutcome
-
-class GradleVerionsParams {
-
-    String gradleVersion = "0"
-
-    boolean expectedGradleBuildFailure = false
-
-    TaskOutcome expectedTaskOutcome = TaskOutcome.FAILED
-
-    @Override
-    String toString() {
-        return "GradleVerionsParams{" +
-                "gradleVersion='" + gradleVersion + '\'' +
-                ", expectedGradleBuildFailure=" + expectedGradleBuildFailure +
-                ", expectedTaskOutcome=" + expectedTaskOutcome +
-                '}'
-    }
-}
