+++ core/src/test/groovy/guru/stefma/bintrayrelease/GradleVerionsParams.groovy
@@ -1,0 +1,21 @@
+package guru.stefma.bintrayrelease
+
+import org.gradle.testkit.runner.TaskOutcome
+
+class GradleVerionsParams {
+
+    String gradleVersion = "0"
+
+    boolean expectedGradleBuildFailure = false
+
+    TaskOutcome expectedTaskOutcome = TaskOutcome.FAILED
+
+    @Override
+    String toString() {
+        return "GradleVerionsParams{" +
+                "gradleVersion='" + gradleVersion + '\'' +
+                ", expectedGradleBuildFailure=" + expectedGradleBuildFailure +
+                ", expectedTaskOutcome=" + expectedTaskOutcome +
+                '}'
+    }
+}
