+++ core/src/test/groovy/guru/stefma/bintrayrelease/TestInvalidExtensionSetup.groovy
@@ -1,0 +1,43 @@
+package guru.stefma.bintrayrelease
+
+import guru.stefma.bintrayrelease.rule.TestProjectRule
+import org.gradle.testkit.runner.GradleRunner
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.junit.runners.JUnit4
+
+import static org.assertj.core.api.Assertions.*
+
+@RunWith(JUnit4.class)
+class TestInvalidExtensionSetup {
+
+    private String buildScript = """
+            plugins { 
+                id 'java-library'
+                id 'guru.stefma.bintrayrelease'
+            }
+            
+            group = "guru.stefma"
+            version = "1.0"
+            publish {
+                userOrg = 'stefma'
+                artifactId = 'test'
+                // desc = 'description'
+            }
+               """
+
+    @Rule
+    public TestProjectRule projectRule = new TestProjectRule(TestProjectRule.Project.JAVA, buildScript)
+
+    @Test
+    void testInvalidExtension_versionAndDescMissing_ShouldFailWithCorrectMessage() {
+        def result = GradleRunner.create()
+                .withProjectDir(projectRule.projectDir)
+                .withArguments("build")
+                .withPluginClasspath()
+                .buildAndFail()
+
+        assertThat(result.output).contains("Have you created the publish closure? Missing desc. ")
+    }
+}
