--- samples/android/src/main/java/com/novoda/gradle/release/sample/android/AndroidSample.java
@@ -1,9 +1,0 @@
-package com.novoda.gradle.release.sample.android;
-
-public class AndroidSample {
-
-    public static void hello() {
-        System.out.println("Hello world from AndroidSample");
-    }
-
-}
