+++ samples/android/src/main/java/guru/stefma/bintrayrelease/sample/android/AndroidSample.java
@@ -1,0 +1,9 @@
+package guru.stefma.bintrayrelease.sample.android;
+
+public class AndroidSample {
+
+    public static void hello() {
+        System.out.println("Hello world from AndroidSample");
+    }
+
+}
