--- samples/jvm/src/main/java/com/novoda/gradle/release/sample/jvm/JVMSample.java
@@ -1,9 +1,0 @@
-package com.novoda.gradle.release.sample.jvm;
-
-public class JVMSample {
-
-    public static void hello() {
-        System.out.println("Hello world from JVMSample");
-    }
-
-}
