+++ samples/jvm/src/main/java/guru/stefma/bintrayrelease/sample/jvm/JVMSample.java
@@ -1,0 +1,9 @@
+package guru.stefma.bintrayrelease.sample.jvm;
+
+public class JVMSample {
+
+    public static void hello() {
+        System.out.println("Hello world from JVMSample");
+    }
+
+}
